requirements.txt
Django==1.9.4
gunicorn==19.6.0
django-tastypie
